remove bufferedsupply, make builder block asserting
