+++ edg/core/Builder.py
@@ -16,10 +16,16 @@ class Builder:
     def __init__(self) -> None:
         self.stack: List[BaseBlock] = []
 
+    def _current_block(self) -> Optional[BaseBlock]:
+        if not self.stack:
+            return None
+        else:
+            return self.stack[-1]
+
     def push_element(self, elt: BaseBlock) -> Optional[BaseBlock]:
         """Pushes a new element onto the context stack, returning the previous top element.
         Ignores if the element is already on top of the stack."""
-        prev_elt = self.get_enclosing_block()
+        prev_elt = self._current_block()
         if not self.stack or self.stack[-1] is not elt:  # prevent double-pushing
             self.stack.append(elt)
         return prev_elt
@@ -33,15 +39,20 @@ class Builder:
         self.stack.pop()
         assert (prev is None and not self.stack) or (prev is not None and self.stack[-1] is prev)
 
-    def get_enclosing_block(self) -> Optional[BaseBlock]:
-        if not self.stack:
-            return None
-        else:
-            return self.stack[-1]
+    def block(self) -> BaseBlock:
+        """Returns the current block context, throwing an error if there is no block context."""
+        current_block = self._current_block()
+        if current_block is None:
+            raise Exception("no block context")
+        return current_block
 
-    @deprecated("use get_enclosing_block() instead, context frames can only be blocks now")
+    @deprecated("use asserting block() instead")
+    def get_enclosing_block(self) -> Optional[BaseBlock]:
+        return self._current_block()
+
+    @deprecated("use block() instead, context frames can only be blocks now")
     def get_curr_context(self) -> Optional[BaseBlock]:
-        return self.get_enclosing_block()
+        return self._current_block()
 
     @overload
     def elaborate_toplevel(
@@ -51,6 +62,7 @@ class Builder:
         is_generator: bool = False,
         generate_values: Iterable[Tuple[edgir.LocalPath, edgir.ValueLit]] = [],
     ) -> edgir.HierarchyBlock: ...
+
     @overload
     def elaborate_toplevel(
         self,
