+++ edg/parts/PowerConditioning.py
@@ -4,134 +4,6 @@ from typing import Optional, cast, Any
 from typing_extensions import override
 
 from ..abstract_parts import *
-from ..electronics_model.PassivePort import PassiveAdapterVoltageSink, PassiveAdapterVoltageSource
-
-
-class Supercap(DiscreteComponent, FootprintBlock):  # TODO actually model supercaps and parts selection
-    def __init__(self) -> None:
-        super().__init__()
-        self.pos = self.Port(VoltageSink())
-        self.neg = self.Port(Ground())
-
-    @override
-    def contents(self) -> None:
-        super().contents()
-        self.footprint(
-            "C",
-            "Capacitor_THT:CP_Radial_D14.0mm_P5.00mm",  # actually 13.5
-            {
-                "1": self.pos,
-                "2": self.neg,
-            },
-            part="DBN-5R5D334T",  # TODO this is too high resistance
-            datasheet="http://www.elna.co.jp/en/capacitor/double_layer/catalog/pdf/dbn_e.pdf",
-        )
-
-
-class BufferedSupply(PowerConditioner):
-    """Implements a current limiting source with an opamp for charging a supercap, and a Vf-limited diode
-    for discharging
-
-    See https://electronics.stackexchange.com/questions/178605/op-amp-mosfet-constant-current-power-source
-    """
-
-    def __init__(self, charging_current: RangeLike, sense_resistance: RangeLike, voltage_drop: RangeLike) -> None:
-        super().__init__()
-
-        self.charging_current = self.ArgParameter(charging_current)
-        self.sense_resistance = self.ArgParameter(sense_resistance)
-        self.voltage_drop = self.ArgParameter(voltage_drop)
-
-        self.pwr = self.Port(VoltageSink.empty(), [Power, Input])
-        self.pwr_out = self.Port(VoltageSource.empty(), [Output])
-        self.require(
-            self.pwr.current_draw.within(
-                self.pwr_out.link().current_drawn + (0, self.charging_current.upper()) + (0, 0.05)
-            )
-        )  # TODO nonhacky bounds on opamp/sense resistor current draw
-        self.sc_out = self.Port(VoltageSource.empty(), optional=True)
-        self.gnd = self.Port(Ground.empty(), [Common])
-
-        max_in_voltage = self.pwr.link().voltage.upper()
-        max_charge_current = self.charging_current.upper()
-
-        # Upstream power domain
-        # TODO improve connect modeling everywhere
-        with self.implicit_connect(
-            ImplicitConnect(self.pwr, [Power]),
-            ImplicitConnect(self.gnd, [Common]),
-        ) as imp:
-            self.sense = self.Block(
-                Resistor(  # TODO replace with SeriesResistor/CurrentSenseResistor - that propagates current
-                    resistance=self.sense_resistance,
-                    power=(0, max_charge_current * max_charge_current * self.sense_resistance.upper()),
-                )
-            )
-            self.connect(self.pwr, self.sense.a.adapt_to(VoltageSink(current_draw=(0, max_charge_current))))
-
-            self.fet = self.Block(
-                Fet.PFet(
-                    drain_voltage=(0, max_in_voltage),
-                    drain_current=(0, max_charge_current),
-                    gate_voltage=(self.pwr.link().voltage.lower(), max_in_voltage),
-                    rds_on=(0, 0.5) * Ohm,  # TODO kind of arbitrary
-                    power=(0, max_in_voltage * max_charge_current),
-                )
-            )
-            self.connect(self.fet.source, self.sense.b)
-
-            self.diode = self.Block(
-                Diode(
-                    reverse_voltage=(0, max_in_voltage),
-                    current=self.charging_current,
-                    voltage_drop=self.voltage_drop,
-                )
-            )
-            self.connect(
-                self.diode.anode.adapt_to(VoltageSink()),
-                self.fet.drain.adapt_to(VoltageSource(voltage_out=self.pwr.link().voltage)),
-                self.sc_out,
-            )
-
-            self.pwr_out_merge = self.Block(MergedVoltageSource()).connected_from(
-                self.pwr,
-                self.diode.cathode.adapt_to(
-                    VoltageSource(
-                        voltage_out=(
-                            self.pwr.link().voltage.lower() - self.voltage_drop.upper(),
-                            self.pwr.link().voltage.upper(),
-                        )
-                    )
-                ),  # TODO replace with SeriesVoltageDiode or something that automatically calculates voltage drops?
-            )
-            self.connect(self.pwr_out_merge.pwr_out, self.pwr_out)
-
-            # TODO check if this tolerance stackup is stacking in the right direction... it might not
-            low_sense_volt_diff = self.charging_current.lower() * self.sense_resistance.lower()
-            high_sense_volt_diff = self.charging_current.upper() * self.sense_resistance.upper()
-            low_sense_volt = self.pwr.link().voltage.lower() - high_sense_volt_diff
-            high_sense_volt = self.pwr.link().voltage.upper() - low_sense_volt_diff
-
-            self.set = imp.Block(
-                VoltageDivider(output_voltage=(low_sense_volt, high_sense_volt), impedance=(1, 10) * kOhm)
-            )
-            self.connect(self.set.input, self.pwr)  # TODO use chain
-            self.amp = imp.Block(Opamp())
-            self.connect(self.set.output, self.amp.inp)
-            self.connect(
-                self.amp.inn,
-                self.sense.b.adapt_to(
-                    AnalogSource(
-                        voltage_out=(0, self.pwr.link().voltage.upper()),
-                        # TODO calculate operating signal level
-                    )
-                ),
-            )
-            self.connect(self.amp.out, self.fet.gate.adapt_to(AnalogSink()))
-
-        self.cap = self.Block(Supercap())
-        self.connect(self.sc_out, self.cap.pos)
-        self.connect(self.gnd, self.cap.neg)
 
 
 class SingleDiodePowerMerge(PowerConditioner, Block):
@@ -295,11 +167,11 @@ class PriorityPowerOr(PowerConditioner, KiCadSchematicBlock, Block):
     ) -> "PriorityPowerOr":
         """Convenience function to connect ports, returning this object so it can still be given a name."""
         if gnd is not None:
-            cast(Block, builder.get_enclosing_block()).connect(gnd, self.gnd)
+            builder.block().connect(gnd, self.gnd)
         if pwr_hi is not None:
-            cast(Block, builder.get_enclosing_block()).connect(pwr_hi, self.pwr_hi)
+            builder.block().connect(pwr_hi, self.pwr_hi)
         if pwr_lo is not None:
-            cast(Block, builder.get_enclosing_block()).connect(pwr_lo, self.pwr_lo)
+            builder.block().connect(pwr_lo, self.pwr_lo)
         return self
