+++ examples/Datalogger/Datalogger.net
@@ -160,7 +160,7 @@
   (value "buffer.sense")
   (footprint "Resistor_SMD:R_0603_1608Metric")
   (property (name "Sheetname") (value "buffer"))
-  (property (name "Sheetfile") (value "edg.parts.PowerConditioning.BufferedSupply"))
+  (property (name "Sheetfile") (value "examples.test_datalogger.BufferedSupply"))
   (property (name "edg_path") (value "buffer.sense"))
   (property (name "edg_short_path") (value "buffer.sense"))
   (property (name "edg_refdes") (value "R5"))
@@ -172,7 +172,7 @@
   (value "buffer.fet")
   (footprint "Package_TO_SOT_SMD:SOT-223-3_TabPin2")
   (property (name "Sheetname") (value "buffer"))
-  (property (name "Sheetfile") (value "edg.parts.PowerConditioning.BufferedSupply"))
+  (property (name "Sheetfile") (value "examples.test_datalogger.BufferedSupply"))
   (property (name "edg_path") (value "buffer.fet"))
   (property (name "edg_short_path") (value "buffer.fet"))
   (property (name "edg_refdes") (value "Q1"))
@@ -184,7 +184,7 @@
   (value "buffer.diode")
   (footprint "Diode_SMD:D_SOD-323")
   (property (name "Sheetname") (value "buffer"))
-  (property (name "Sheetfile") (value "edg.parts.PowerConditioning.BufferedSupply"))
+  (property (name "Sheetfile") (value "examples.test_datalogger.BufferedSupply"))
   (property (name "edg_path") (value "buffer.diode"))
   (property (name "edg_short_path") (value "buffer.diode"))
   (property (name "edg_refdes") (value "D1"))
@@ -244,7 +244,7 @@
   (value "buffer.cap")
   (footprint "Capacitor_THT:CP_Radial_D14.0mm_P5.00mm")
   (property (name "Sheetname") (value "buffer"))
-  (property (name "Sheetfile") (value "edg.parts.PowerConditioning.BufferedSupply"))
+  (property (name "Sheetfile") (value "examples.test_datalogger.BufferedSupply"))
   (property (name "edg_path") (value "buffer.cap"))
   (property (name "edg_short_path") (value "buffer.cap"))
   (property (name "edg_refdes") (value "C6"))
